fix (1) try with newline (2) move methods from abstract to test
